+++ testable-simd-models/src/abstractions/funarr.rs
@@ -5,19 +5,19 @@
 /// Internally, it uses a fixed-length array of `Option<T>` with a maximum capacity of 512 elements.
 /// Unused elements beyond `N` are filled with `None`.
 #[derive(Copy, Clone, Eq, PartialEq)]
-pub struct FunArray<const N: u64, T>([Option<T>; 512]);
+pub struct FunArray<const N: u32, T>([Option<T>; 512]);
 
-impl<const N: u64, T> FunArray<N, T> {
+impl<const N: u32, T> FunArray<N, T> {
     /// Gets a reference to the element at index `i`.
-    pub fn get(&self, i: u64) -> &T {
+    pub fn get(&self, i: u32) -> &T {
         self.0[i as usize].as_ref().unwrap()
     }
     /// Constructor for FunArray. `FunArray<N,T>::from_fn` constructs a funarray out of a function that takes usizes smaller than `N` and produces an element of type T.
-    pub fn from_fn<F: Fn(u64) -> T>(f: F) -> Self {
+    pub fn from_fn<F: Fn(u32) -> T>(f: F) -> Self {
         // let vec = (0..N).map(f).collect();
         let arr = core::array::from_fn(|i| {
-            if (i as u64) < N {
-                Some(f(i as u64))
+            if (i as u32) < N {
+                Some(f(i as u32))
             } else {
                 None
             }
@@ -53,10 +53,10 @@ impl<const N: u64, T> FunArray<N, T> {
     }
 }
 
-impl<const N: u64, T: Clone> TryFrom<Vec<T>> for FunArray<N, T> {
+impl<const N: u32, T: Clone> TryFrom<Vec<T>> for FunArray<N, T> {
     type Error = ();
     fn try_from(v: Vec<T>) -> Result<Self, ()> {
-        if (v.len() as u64) < N {
+        if (v.len() as u32) < N {
             Err(())
         } else {
             Ok(Self::from_fn(|i| v[i as usize].clone()))
@@ -64,16 +64,16 @@ impl<const N: u64, T: Clone> TryFrom<Vec<T>> for FunArray<N, T> {
     }
 }
 
-impl<const N: u64, T: core::fmt::Debug + Clone> core::fmt::Debug for FunArray<N, T> {
+impl<const N: u32, T: core::fmt::Debug + Clone> core::fmt::Debug for FunArray<N, T> {
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{:?}", self.as_vec())
     }
 }
 
-impl<const N: u64, T> core::ops::Index<u64> for FunArray<N, T> {
+impl<const N: u32, T> core::ops::Index<u32> for FunArray<N, T> {
     type Output = T;
 
-    fn index(&self, index: u64) -> &Self::Output {
+    fn index(&self, index: u32) -> &Self::Output {
         self.get(index)
     }
 }
